project: Integrate native IPC version of obs-studio-node
This version of obs-studio-node moves the libOBS (and dependencies, plugins, etc) out of process using a custom IPC solution that can handle lots of data without having to block the main process. With this the main process and renderer process can load anything they want without having any impact on the streaming and recording and it also allows us to restart the UI in the event of a crash and keep almost everything working - streamers no longer have to fear an electron/UI crash, as we should be able to recover from that.

In the future, it might be possible for renderer processes to directly call obs-studio-node instead of going through the main process, which saves additional cycles and frees up even more resources, and also lowers CPU usage required by electron.